pyproject.toml
[project]
name = "unifor-projeto"
version = "0.1.0"
description = ""
authors = [
    {name = "Lucas Mesquita",email = "pr1v4t3.lm@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.10"
dependencies = [
    "requests (==2.31.0)",
    "pandas (==2.2.2)",
    "streamlit (==1.35.0)",
    "pymongo (==4.7.2)",
    "kafka-python (>=2.2.15,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "watermark (>=2.5.0,<3.0.0)",
    "openai (==0.28)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"

